change: Use TypeScript for commands testing module
